Extract link re-checking into LinkCheckingConcern
LinksController and CitationsController had near-identical synchronous
link-check logic (and an identical link_toast_for). Move both into a shared
LinkCheckingConcern; callers pass the record, the URL, and the status/checked_at
column names (which differ between LexLink and LexCitation). Behavior is
unchanged, including the blank-URL clearing path and the toast flash.
