+++ app/controllers/lexicon/links_controller.rb
@@ -4,6 +4,7 @@ module Lexicon
   # Controller to work with Lexicon Links
   class LinksController < ApplicationController
     include LockLexEntryConcern
+    include LinkCheckingConcern
 
     before_action do
       require_editor('edit_lexicon')
@@ -36,7 +37,9 @@ module Lexicon
       if @link.update(lex_link_params)
         # Re-check the link only when its URL actually changed, so a previously-broken
         # link (e.g. HTTP 403) is re-evaluated instead of keeping its stale status.
-        check_link_synchronously if @link.saved_change_to_url?
+        if @link.saved_change_to_url?
+          check_link_synchronously(@link, @link.url, status_column: :http_status, checked_at_column: :checked_at)
+        end
         return
       end
 
@@ -67,30 +70,6 @@ module Lexicon
     # Only allow a list of trusted parameters through.
     def lex_link_params
       params.expect(lex_link: %i(url description))
-    end
-
-    # Re-checks the (changed) URL synchronously and stores the fresh status, mirroring
-    # Lexicon::CitationsController#check_link_synchronously. url has a presence validation,
-    # so it is never blank here. The toast is stashed in flash for the JS response.
-    def check_link_synchronously
-      status = CheckExternalLinks.new.check_url(@link.url)
-      @link.update_columns(http_status: status, checked_at: Time.current)
-      @link_check_performed = true
-      @link_toast_type, @link_toast_message = link_toast_for(status)
-      # rubocop:disable Rails/ActionControllerFlashBeforeRender
-      flash[:link_check_toast_type] = @link_toast_type
-      flash[:link_check_toast_message] = @link_toast_message
-      # rubocop:enable Rails/ActionControllerFlashBeforeRender
-    end
-
-    def link_toast_for(status)
-      if status.nil?
-        ['error', t('lexicon.verification.broken_link.inaccessible')]
-      elsif status < 400
-        ['success', t('lexicon.verification.broken_link.now_accessible')]
-      else
-        ['error', t('lexicon.verification.broken_link.still_broken', status: status)]
-      end
     end
   end
 end
